+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 0.1.9 (unreleased)
+* Update experimental in-eval ref lookup. Now exposed by two
+  functions, `bml.ref(id)` (returns rendered output of ref) and
+  `bml.refDetail(id)` (returns ref output, index, and original fork
+  object). The new functions should be lazily evaluated so they act on
+  the fork map state at the time of function use, not function
+  declaration.
+
 ### 0.1.8
 * Support set forks with `{$id: (foo), (bar)}` syntax. Silent set
   forks are also supported.